rename: 폴더 구조 변경 - pages 폴더 생성, container 컴포넌트 분리
