+++ src/pages/shareFolder/index.js
@@ -1,0 +1,5 @@
+import Share from "components/units/shareFolder/Share.container.js";
+
+export default function SharedPage() {
+  return <Share />;
+}
